+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 * Added a game variant message at the start of a game. ([#235](https://github.com/hjwylde/werewolf/issues/235))
 * Added the Saint role. ([#178](https://github.com/hjwylde/werewolf/issues/178))
 * Added marks to the Dullahan's status command. ([#236](https://github.com/hjwylde/werewolf/issues/236))
+* Changed Spiteful Ghost to be aligned with the Villagers and to start off alive. ([#238](https://github.com/hjwylde/werewolf/issues/238))
 
 *Revisions*